User sessioning enabled in the Groceries and Communities spaces
When not logged, the user is asked to log in and redirected to the login
screen when desired. There, he logs in and is redirected back to the
previous screen he was trying to access.

Each space's resources are displayed/loaded/modified/stored depending on
the user's authentication status and expertise status.

Next, working on adding new post in a community